moved IEthereumContext to where it is used
